Update main.dart
This is a temp main.dart. 
The home page displays all the trips registered. 
For each trip, it shows the destination, start and end date. To view the remaining details, the user need to click on view more. 
At the bottom right, there is a + icon, which once clicked, the form will show up. The user will then enter all required trip info and register a new trip.